+++ src/methods/setDefaultPaymentMethod.ts
@@ -3,16 +3,16 @@ import { stripeApiUrl, stripeApiVersion } from "../utils/constants";
 import { RemedyProductStripe } from ".";
 
 /**
- * Set default customer card
+ * Set default customer payment method
  *
- * @param cardId - card id (see: https://stripe.com/docs/api/customers/object#card_object-id)
+ * @param paymentMethodId - payment method id (see: https://stripe.com/docs/api/customers/object#payment_method_object-id)
  * @param customerId - customer id (see: https://stripe.com/docs/api/customers/object#customer_object-id)
  * @param ephemeralKey - customer ephemeral key
  * @returns
  */
-export const setDefaultCard = async function (
+export const setDefaultPaymentMethod = async function (
   this: RemedyProductStripe,
-  cardId: string,
+  paymentMethodId: string,
   customerId: string,
   ephemeralKey: string
 ) {
@@ -21,10 +21,9 @@ export const setDefaultCard = async function (
   /* eslint-enable */
   if (typeof stripeApiKey !== "string")
     throw new Error("Initialization failed.");
-
-  // make request
+  // make request for payment method api
   return fetch(`${stripeApiUrl}/customers/${customerId}`, {
-    body: `default_source=${cardId}`,
+    body: `invoice_settings[default_payment_method]=${paymentMethodId}`,
     headers: {
       Authorization: `Bearer ${ephemeralKey}`,
       "Content-Type": "application/x-www-form-urlencoded",