+++ CHANGELOG.md
@@ -1,9 +1,24 @@
+### v1.0.0
+
+- Updated libs
+- Added method `addPaymentMethodToCustomer`
+- Added method `confirmPaymentIntentByPaymentMethod`
+- Added method `deletePaymentMethodFromCustomer`
+- Added method `getAllPaymentMethods`
+- Added method `getDefaultCard`
+- Added method `setDefaultPaymentMethod`
+- Changed method `getAllCards`
+- Changed method `getCustomer`
+- Changed method `setDefaultCard`
+
 ### v0.0.6
 
 - Add type's declaration
+
 ### v0.0.5
 
 - Add new param to `confirmPaymentIntentByCard`
+
 ### v0.0.4
 
 - Added method `getAllCards`